\name{rasterDT-package}
\alias{rasterDT-package}
\alias{rasterDT}
\docType{package}
\title{\packageTitle{rasterDT}}
\description{\packageDescription{rasterDT}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{rasterDT}
  \packageIndices{rasterDT}

  Fast alternatives to several relatively slow raster package
  functions. For large rasters, the functions run from 5 to
  approximately 100 times faster than the raster package functions
  they replace.
}
\author{
  \packageAuthor{rasterDT}
  
  Maintainer: \packageMaintainer{rasterDT}
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
