% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstabDT.R
\name{crosstabDT}
\alias{crosstabDT}
\title{Speedy Cross-tabulation of Two Categorical Rasters}
\usage{
crosstabDT(x, y, digits = 0, long = FALSE, useNA = FALSE)
}
\arguments{
\item{x}{A \code{RasterLayer} whose values give the index of the
\code{POLYGON} or \code{MULTIPOLYGON} fragment from which each
cell was rasterized.}

\item{y}{A categorical \code{RasterLayer} whose values indicate
the habitat classification of the area represented by each
cell.}
}
\value{
A \code{data.table} recording the frequency of each
    habitat type within each indexed area. The first column,
    \code{"x"}, gives an index corresponding to the \code{"x"}
    column of the \code{sp} layer object from which the \code{x}
    raster was rasterized.
}
\description{
A fast data.table-based drop-in replacement for
\code{\link[raster:crosstab]{raster::crosstab()}}.
}
\examples{
r <- raster(nc = 5, nr = 5)
r[] <- sample(5, 25, replace = TRUE)
s <- setValues(r, sample(2, 25, replace = TRUE))
crosstabDT(r,s)
}
\author{
Joshua O'Brien
}
