% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterizeDT.R
\name{fasterizeDT}
\alias{fasterizeDT}
\title{Fixed up fasterize Function}
\usage{
fasterizeDT(x, raster, field = NULL, fun = "last",
  background = NA_real_, by = NULL)
}
\arguments{
\item{x}{An \code{sf::sf()} object with a geometry column of
\code{POLYGON} and/or \code{MULTIPOLYGON} objects or a
\code{sp::SpatialPolygonsDataFrame} object.}

\item{raster}{A raster object. Used as a template for the raster
output}

\item{field}{character. The name of a column in \code{x},
providing a value for each of the polygons rasterized. If NULL
(default), all polygons will be given a value of 1.}

\item{fun}{\code{?fasterize::fasterize}}

\item{background}{Value to put in the cells that are not covered
by any of the features of x. Default is NA.}

\item{by}{\code{?fasterize::fasterize}}
}
\value{
A raster of the same size, extent, resolution and
    projection as the supplied raster template. Unlike
    \code{\link[fasterize:fasterize]{fasterize::fasterize()}},
    \code{fasterizeDT} returns a raster of the same type as the
    data in the column of \code{x} column selected by the
    \code{field} argument.
}
\description{
A front end for
\code{\link[fasterize:fasterize]{fasterize::fasterize()}}, fixing
several of its infelicities.
}
\examples{
SPDF <- shapefile(system.file("external/lux.shp", package = "raster"))
llratio <- 1/cos(pi * mean(coordinates(SPDF)[,2])/180)
rr <- raster(extent(SPDF),
             resolution = c(llratio * 0.01, 0.01),
             crs = proj4string(SPDF))

## An integer-valued field produces a numeric raster
rInt <- fasterizeDT(SPDF, rr, field = "ID_2")
plot(rInt, col = colorRampPalette(blues9)(12))

## A character-valued field returns a factor raster
rFac <- fasterizeDT(SPDF, rr, field = "NAME_2")
if (require(rasterVis)) {
    levelplot(rFac)
}
}
\author{
Joshua O'Brien
}
