% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsDT.R
\name{subsDT}
\alias{subsDT}
\title{Speedy Substitution of Values in a RasterLayer}
\usage{
subsDT(x, dict, by = 1, which = 2, subsWithNA = TRUE,
  filename = "", ...)
}
\arguments{
\item{x}{Categorical \code{RasterLayer} with integer values giving
field class.}

\item{dict}{A \code{data.frame} or \code{data.table} with one
column corresponding to the levels of cells in \code{x} and
another column giving the value to which each level should be
mapped.}

\item{by}{Character string giving the name of the column in
\code{dict} containing the categorical values in \code{x}.}

\item{which}{Character string giving name of the column in
\code{dict} with the numerical values to which each value in
\code{by} should be mapped.}

\item{subsWithNA}{Logical. If 'TRUE' values that are not matched
become NA.  If 'FALSE', they retain their original value
(which could also be 'NA'). This latter option is handy when
you want to replace only one or a few values. It cannot be
used when 'x' has multiple layers}

\item{filename}{Character string giving (optional) file name.}

\item{...}{Additional arguments as for
\code{\link[raster:writeRaster]{raster::writeRaster()}}, on
which this function relies.}
}
\value{
A \code{RasterLayer} object.
}
\description{
A fast \code{data.table}-based drop-in replacement for
\code{\link[raster:substitute]{raster::subs()}}.
}
\examples{
## Small example

r <- raster(ncol = 10, nrow = 10)
vals <- rep(1:5, each = 20)
r <- setValues(r, vals)
set.seed(4)
dict <- data.table(ID = 1:5, VAL = rnorm(5))
out <- subsDT(r, dict)
plot(out)
}
\author{
Joshua O'Brien
}
