% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsDT.R
\name{subsDT}
\alias{subsDT}
\title{Speedy Substitution of Values in a RasterLayer}
\usage{
subsDT(r, dict, by = "ID", which = "VAL", filename = "", ...)
}
\arguments{
\item{r}{Categorical \code{RasterLayer} with integer values giving
field class.}

\item{dict}{A \code{data.frame} or \code{data.table} with one
column corresponding to the levels of cells in \code{r} and
another column giving the value to which each level should be
mapped.}

\item{by}{Character string giving the name of the column in
\code{dict} containing the categorical values in \code{r}.}

\item{which}{Character string giving name of the column in
\code{dict} with the numerical values to which each value in
\code{by} should be mapped.}

\item{filename}{Character string giving (optional) file name.}

\item{...}{Additional arguments as for
\code{\link[raster:writeRaster]{raster::writeRaster()}}, on
which this function relies.}
}
\value{
A \code{RasterLayer} object.
}
\description{
A fast \code{data.table}-based drop-in replacement for
\code{\link[raster:substitute]{raster::subs()}}.
}
\examples{
## Small example

r <- raster(ncol = 10, nrow = 10)
vals <- rep(1:5, each = 20)
r <- setValues(r, vals)
set.seed(4)
dict <- data.table(ID = 1:5, VAL = rnorm(5))
out <- subsDT(r, dict)
plot(out)
}
\author{
Joshua O'Brien
}
